pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.3",
    "uvicorn>=0.37.0",
]
